disable eval dataloader replacement
